Use styles.js
- 공통 CSS가 있다면 styles.js 로 묶고 가져다 사용하자
