--- src/MenuList.jsx
@@ -1,10 +1,0 @@
-import styled from '@emotion/styled';
-
-const MenuList = styled.ul({
-  display: 'flex',
-  listStyle: 'none',
-  margin: 0,
-  padding: '.4em 0',
-});
-
-export default MenuList;
